feat(impl): one shared <Markdown> renderer makes the Tasks timeline + detail readable [conductor:f01b6619]
